// src/types/ganttTaskActions.ts
import { BarTask } from './barTask';

export type BarMoveAction = 'progress' | 'end' | 'start' | 'move';
export type GanttContentMoveAction =
  | 'mouseenter'
  | 'mouseleave'
  | 'delete'
  | 'dblclick'
  | 'select'
  | ''
  | BarMoveAction;

export type GanttEvent = {
  changedTask?: BarTask;
  originalSelectedTask?: BarTask;
  action: GanttContentMoveAction;
};